backend/requirements.txt
fastapi
uvicorn
asyncpg
pydantic
python-jose
passlib
python-multipart
python-dotenv